Berry more automated solidification
