+++ lib/libesp32/berry_tasmota/src/solidify/solidified_partition_core.h
@@ -1,9 +1,7 @@
-/********************************************************************
- * Berry module `partition_core`
- * 
- * To use: `import partition_core`
- * 
- *******************************************************************/
+/* Solidification of partition_core.h */
+/********************************************************************\
+* Generated code, don't edit                                         *
+\********************************************************************/
 #include "be_constobj.h"
 
 /********************************************************************
